Checked old exercises and add new

+++ scripts/ft_strlen.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char* str)
+size_t	ft_strlen(const char* str)
 {
 	int	i;
